backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
xgboost
pandas
geopy
joblib
python-multipart
requests
python-dotenv
Pillow
earthaccess
xarray
h5netcdf